ai_principal/mcp_server/requirements.txt
# MCP server SDK (ajuste o nome conforme a SDK escolhida)
# Cliente Supabase para Python
supabase>=2.0.0

# Utilitários HTTP (opcional)
httpx>=0.25.0

# Variáveis de ambiente
python-dotenv>=1.0.0

# (Opcional) FastAPI para endpoints HTTP
fastapi>=0.110.0
uvicorn>=0.29.0

# (Opcional) Logging estruturado
loguru>=0.7.0
